Change signature of dispatchEvent to take in LogEvent object rather than url and request parameters
