+++ core-api/src/main/java/com/optimizely/ab/event/LogEvent.java
